Merge branch 'main' of https://github.com/RedbirdXVO/CharacterCreatorProject
Made adjustments to the following:

Adjusted existing buttons to use Vertical and horizontal grid patterns (For ease of adding further UI - I have a backup of the old file if we find this is too hard to work with)

Created custom button assets and background image. Using slicing, we are able to adjust button/background size without having to create multiple images.

Finally, instated a RGB slider for skintone. We can absolutely go with set options, if we wanted a more realistic approach. But the RGB slider is there if we need it.